Move input binding defaults into config crate
This keeps all the input definitions in one spot, which is less confusing.
